% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycle1}
\alias{cycle1}
\title{Canadian Health Measures Survey (CHMS) Cycle 1}
\source{
Statistics Canada
}
\usage{
data(cycle1)
}
\description{
This is dummy data representing the second cycle of the Canadian
Health Measures Survey (CHMS). The CHMS survey is conducted by
Statistics Canada.
}
\examples{
data(cycle1)
str(cycle1)
}
\keyword{datasets}
