% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycle1_meds}
\alias{cycle1_meds}
\title{Canadian Health Measures Survey (CHMS) Cycle 1 Medications}
\format{
A data frame with X rows and Y columns.
}
\source{
Statistics Canada
}
\usage{
data(cycle1_meds)
}
\description{
This dummy data representing the medication portion of the
second cycle of the Canadian Health Measures Survey (CHMS).
The CHMS survey is conducted by Statistics Canada.
}
\examples{
data(cycle1_meds)
str(cycle1_meds)
}
\keyword{datasets}
