% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cycle2}
\alias{cycle2}
\title{CHMS Cycle 2}
\description{
This is synthetic data representing the second cycle of the Canadian
Health Measures Survey (CHMS). The CHMS survey is conducted by
Statistics Canada.
}
