% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cycle3}
\alias{cycle3}
\title{CHMS Cycle 3}
\description{
This is synthetic data representing the third cycle of the Canadian
Health Measures Survey (CHMS). The CHMS survey is conducted by
Statistics Canada.
}
