% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cycle3_meds}
\alias{cycle3_meds}
\title{CHMS Cycle 3 Medications}
\description{
This synthetic data representing the medication portion of the
third cycle of the Canadian Health Measures Survey (CHMS).
The CHMS survey is conducted by Statistics Canada.
}
