% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cycle4}
\alias{cycle4}
\title{CHMS Cycle 4}
\description{
This is synthetic data representing the fourth cycle of the Canadian
Health Measures Survey (CHMS). The CHMS survey is conducted by
Statistics Canada.
}
