% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycle5}
\alias{cycle5}
\title{Canadian Health Measures Survey (CHMS) Cycle 5}
\format{
A data frame with X rows and Y columns.
}
\source{
Statistics Canada
}
\usage{
data(cycle5)
}
\description{
This is dummy data representing the fifth cycle of the Canadian
Health Measures Survey (CHMS). The CHMS survey is conducted by
Statistics Canada.
}
\examples{
data(cycle5)
str(cycle5)
}
\keyword{datasets}
