% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycle6}
\alias{cycle6}
\title{Canadian Health Measures Survey (CHMS) Cycle 6}
\source{
Statistics Canada
}
\usage{
data(cycle6)
}
\description{
This is dummy data representing the fifth cycle of the Canadian
Health Measures Survey (CHMS). The CHMS survey is conducted by
Statistics Canada.
}
\examples{
data(cycle6)
str(cycle6)
}
\keyword{datasets}
