% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycle6_meds}
\alias{cycle6_meds}
\title{Canadian Health Measures Survey (CHMS) Cycle 6 Medications}
\source{
Statistics Canada
}
\usage{
data(cycle6_meds)
}
\description{
This dummy data representing the medication portion of the
third cycle of the Canadian Health Measures Survey (CHMS).
The CHMS survey is conducted by Statistics Canada.
}
\examples{
data(cycle6_meds)
str(cycle6_meds)
}
\keyword{datasets}
