% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medications.R
\name{cycles1to2_beta_blockers}
\alias{cycles1to2_beta_blockers}
\title{Beta blockers - cycles 1-2}
\usage{
cycles1to2_beta_blockers(...)
}
\arguments{
\item{...}{Medication and last taken variables.}
}
\value{
\link{numeric} Returns 1 if the respondent is taking beta blockers, 0 otherwise. If all medication information is missing, returns a tagged NA.
}
\description{
This function checks if a person is taking beta blockers based on the provided Anatomical Therapeutic Chemical (ATC) codes for medications
and the Canadian Health Measures Survey (CHMS) response for the time when the medication was last taken.
}
\details{
The function identifies beta blockers based on ATC codes starting with "C07", excluding specific sub-codes. It checks all medication variables provided in the input data frame.

\if{html}{\out{<div class="sourceCode">}}\preformatted{     **Missing Data Codes:**
     - The function handles tagged NAs from the `is_beta_blocker` function and propagates them.
}\if{html}{\out{</div>}}
}
\examples{
# This is a wrapper function and is not intended to be called directly by the user.
# See `is_beta_blocker` for usage examples.
}
\seealso{
\code{is_beta_blocker}
}
