% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medications.R
\name{cycles1to2_calcium_channel_blockers}
\alias{cycles1to2_calcium_channel_blockers}
\title{Calcium channel blockers - cycles 1-2}
\usage{
cycles1to2_calcium_channel_blockers(...)
}
\arguments{
\item{...}{Medication and last taken variables.}
}
\value{
\link{numeric} Returns 1 if the person is taking calcium channel blockers, 0 otherwise. If all medication information is missing, it returns a tagged NA.
}
\description{
This function checks if a person is taking calcium channel blockers based on the provided Anatomical Therapeutic Chemical (ATC) codes for medications
and the Canadian Health Measures Survey (CHMS) response for the time when the medication was last taken.
}
\details{
The function identifies calcium channel blockers based on ATC codes starting with "C08". It checks all medication variables provided in the input data frame.

\if{html}{\out{<div class="sourceCode">}}\preformatted{     **Missing Data Codes:**
     - The function handles tagged NAs from the `is_calcium_channel_blocker` function and propagates them.
}\if{html}{\out{</div>}}
}
\examples{
# This is a wrapper function and is not intended to be called directly by the user.
# See `is_calcium_channel_blocker` for usage examples.
}
\seealso{
\code{is_calcium_channel_blocker}
}
