% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood-pressure.R
\name{determine_adjusted_hypertension}
\alias{determine_adjusted_hypertension}
\title{Determine Adjusted Hypertension Status}
\usage{
determine_adjusted_hypertension(
  SBP_adj,
  DBP_adj,
  ANYMED2,
  CCC_32 = 2,
  CARDIOV = 2,
  DIABX = 2,
  CKD = 2
)
}
\arguments{
\item{SBP_adj}{An integer representing the adjusted systolic blood pressure measurement of the respondent.}

\item{DBP_adj}{An integer representing the adjusted diastolic blood pressure measurement of the respondent.}

\item{ANYMED2}{An integer indicating whether the respondent is on medication for hypertension.
\itemize{
\item 1: Yes
\item 0: No
}}

\item{CCC_32}{An optional integer indicating whether the respondent is actually on medication for hypertension.
\itemize{
\item 1: Yes
\item 2: No (default)
}}

\item{CARDIOV}{An optional integer indicating the presence of cardiovascular disease, affecting medication status.
\itemize{
\item 1: Yes
\item 2: No (default)
}}

\item{DIABX}{An optional integer indicating the presence of diabetes, affecting blood pressure thresholds.
\itemize{
\item 1: Yes
\item 2: No (default)
}}

\item{CKD}{An optional integer indicating the presence of chronic kidney disease, affecting blood pressure thresholds.
\itemize{
\item 1: Yes
\item 2: No (default)
}}
}
\value{
An integer representing the hypertension status:
\itemize{
\item 1: High blood pressure (adjusted BP ≥ 140/90 mmHg (or ≥ 130/80 mmHg if diabetes or CKD) or on hypertension medication)
\item 2: Normal blood pressure (adjusted BP < 140/90 mmHg (or < 130/80 mmHg if diabetes or CKD) and not on hypertension medication)
\item NA(b): Invalid input or non-response
}
}
\description{
This function determines the adjusted hypertension status of a respondent based on their adjusted systolic and diastolic blood pressure measurements and medication usage.
}
\examples{

# Example 1: Respondent has adjusted SBP = 150, adjusted DBP = 95, and on medication.
determine_adjusted_hypertension(SBP_adj = 150, DBP_adj = 95, ANYMED2 = 1)
# Output: 1 (High blood pressure due to adjusted SBP, adjusted DBP, and medication usage).

# Example 2: Respondent has adjusted SBP = 120, adjusted DBP = 80, and not on medication.
determine_adjusted_hypertension(SBP_adj = 120, DBP_adj = 80, ANYMED2 = 2)
# Output: 2 (Normal blood pressure as adjusted BP is below 140/90 mmHg and not on medication).

}
