% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood-pressure.R
\name{determine_controlled_hypertension}
\alias{determine_controlled_hypertension}
\title{Determine Controlled Hypertension Status}
\usage{
determine_controlled_hypertension(
  BPMDPBPS,
  BPMDPBPD,
  ANYMED2,
  CCC_32 = 2,
  CARDIOV = 2,
  DIABX = 2,
  CKD = 2
)
}
\arguments{
\item{BPMDPBPS}{An integer representing the systolic blood pressure measurement of the respondent.}

\item{BPMDPBPD}{An integer representing the diastolic blood pressure measurement of the respondent.}

\item{ANYMED2}{An integer indicating whether the respondent is on medication for hypertension.
\itemize{
\item 1: Yes
\item 0: No
}}

\item{CCC_32}{An optional integer indicating whether the respondent is actually on medication for hypertension.
\itemize{
\item 1: Yes
\item 2: No (default)
}}

\item{CARDIOV}{An optional integer indicating the presence of cardiovascular disease, affecting medication status.
\itemize{
\item 1: Yes
\item 2: No (default)
}}

\item{DIABX}{An optional integer indicating the presence of diabetes, affecting blood pressure thresholds.
\itemize{
\item 1: Yes
\item 2: No (default)
}}

\item{CKD}{An optional integer indicating the presence of chronic kidney disease, affecting blood pressure thresholds.
\itemize{
\item 1: Yes
\item 2: No (default)
}}
}
\value{
An integer representing the hypertension status:
\itemize{
\item 1: Hypertension controlled (BP < 140/90 mmHg (or < 130/80 mmHg if diabetes or CKD) when on hypertension medication)
\item 2: Hypertension not controlled (BP ≥ 140/90 mmHg (or ≥ 130/80 mmHg if diabetes or CKD) when on hypertension medication)
\item NA(b): Invalid input or non-response
}
}
\description{
This function determines the controlled hypertension status of a respondent based on their systolic and diastolic blood pressure measurements and medication usage.
}
\examples{

# Example 1: Determine controlled hypertension status for a respondent with systolic BP = 150, diastolic BP = 95, and on medication.
determine_controlled_hypertension(BPMDPBPS = 150, BPMDPBPD = 95, ANYMED2 = 1)
# Output: 2 (Hypertension not controlled due to high systolic and diastolic BP despite medication usage).

# Example 2: Determine controlled hypertension status for a respondent with systolic BP = 120, diastolic BP = 80, and on medication.
determine_controlled_hypertension(BPMDPBPS = 120, BPMDPBPD = 80, ANYMED2 = 1)
# Output: 1 (Hypertension controlled as BP is below 140/90 mmHg and on medication).

}
