% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-history.R
\name{determine_CVD_personal_history}
\alias{determine_CVD_personal_history}
\title{Cardiovascular disease (CVD) personal history}
\usage{
determine_CVD_personal_history(CCC_61, CCC_63, CCC_81)
}
\arguments{
\item{CCC_61}{\link{integer} An integer representing the respondent's personal history of heart disease. 1 for "Yes" if the person has
heart disease, 2 for "No" if the person does not have heart disease.}

\item{CCC_63}{\link{integer} An integer representing the respondent's personal history of heart attack. 1 for "Yes" if the person had
a heart attack, 2 for "No" if the person did not have a heart attack.}

\item{CCC_81}{\link{integer} An integer representing the respondent's personal history of stroke. 1 for "Yes" if the person had a stroke,
2 for "No" if the person did not have a stroke.}
}
\value{
\link{integer} The CVD personal history: 1 for "Yes" if the person had heart disease, heart attack,
or stroke; 2 for "No" if the person had neither of the conditions; and NA if all the input variables are a
non-response.
}
\description{
This function determines a respondent's cardiovascular disease (CVD) personal history based on the presence or absence
of specific conditions related to heart disease, heart attack, and stroke.
}
\examples{
# Scalar usage: Single respondent
# Determine CVD personal history for a person with heart disease (CCC_61 = 1).
determine_CVD_personal_history(CCC_61 = 1, CCC_63 = 2, CCC_81 = 2)
# Output: 1 (CVD personal history is "Yes" as heart disease is present).

# Example: Respondent has non-response values for all inputs.
determine_CVD_personal_history(CCC_61 = 8, CCC_63 = 8, CCC_81 = 8)
# Output: NA

# Multiple respondents
determine_CVD_personal_history(CCC_61 = c(1, 2, 2), CCC_63 = c(2, 1, 2), CCC_81 = c(2, 2, 1))
# Returns: c(1, 1, 1)

# Database usage: Applied to survey datasets
library(dplyr)
# dataset \%>\%
#   mutate(cvd_personal_history = determine_CVD_personal_history(CCC_61, CCC_63, CCC_81))

}
