% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medications.R
\name{is_ccb_med_cycles1to2}
\alias{is_ccb_med_cycles1to2}
\title{Calcium channel blockers - cycles 1-2}
\usage{
is_ccb_med_cycles1to2(
  atc_101a = NULL,
  atc_102a = NULL,
  atc_103a = NULL,
  atc_104a = NULL,
  atc_105a = NULL,
  atc_106a = NULL,
  atc_107a = NULL,
  atc_108a = NULL,
  atc_109a = NULL,
  atc_110a = NULL,
  atc_111a = NULL,
  atc_112a = NULL,
  atc_113a = NULL,
  atc_114a = NULL,
  atc_115a = NULL,
  atc_201a = NULL,
  atc_202a = NULL,
  atc_203a = NULL,
  atc_204a = NULL,
  atc_205a = NULL,
  atc_206a = NULL,
  atc_207a = NULL,
  atc_208a = NULL,
  atc_209a = NULL,
  atc_210a = NULL,
  atc_211a = NULL,
  atc_212a = NULL,
  atc_213a = NULL,
  atc_214a = NULL,
  atc_215a = NULL,
  atc_131a = NULL,
  atc_132a = NULL,
  atc_133a = NULL,
  atc_134a = NULL,
  atc_135a = NULL,
  atc_231a = NULL,
  atc_232a = NULL,
  atc_233a = NULL,
  atc_234a = NULL,
  atc_235a = NULL,
  mhr_101b = NULL,
  mhr_102b = NULL,
  mhr_103b = NULL,
  mhr_104b = NULL,
  mhr_105b = NULL,
  mhr_106b = NULL,
  mhr_107b = NULL,
  mhr_108b = NULL,
  mhr_109b = NULL,
  mhr_110b = NULL,
  mhr_111b = NULL,
  mhr_112b = NULL,
  mhr_113b = NULL,
  mhr_114b = NULL,
  mhr_115b = NULL,
  mhr_201b = NULL,
  mhr_202b = NULL,
  mhr_203b = NULL,
  mhr_204b = NULL,
  mhr_205b = NULL,
  mhr_206b = NULL,
  mhr_207b = NULL,
  mhr_208b = NULL,
  mhr_209b = NULL,
  mhr_210b = NULL,
  mhr_211b = NULL,
  mhr_212b = NULL,
  mhr_213b = NULL,
  mhr_214b = NULL,
  mhr_215b = NULL,
  mhr_131b = NULL,
  mhr_132b = NULL,
  mhr_133b = NULL,
  mhr_134b = NULL,
  mhr_135b = NULL,
  mhr_231b = NULL,
  mhr_232b = NULL,
  mhr_233b = NULL,
  mhr_234b = NULL,
  mhr_235b = NULL
)
}
\arguments{
\item{atc_101a}{\link{character} ATC code of respondent's first prescription medication.}

\item{atc_102a}{\link{character} ATC code of respondent's second prescription medication.}

\item{atc_103a}{\link{character} ATC code of respondent's third prescription medication.}

\item{atc_104a}{\link{character} ATC code of respondent's fourth prescription medication.}

\item{atc_105a}{\link{character} ATC code of respondent's fifth prescription medication.}

\item{atc_106a}{\link{character} ATC code of respondent's sixth prescription medication.}

\item{atc_107a}{\link{character} ATC code of respondent's seventh prescription medication.}

\item{atc_108a}{\link{character} ATC code of respondent's eighth prescription medication.}

\item{atc_109a}{\link{character} ATC code of respondent's ninth prescription medication.}

\item{atc_110a}{\link{character} ATC code of respondent's tenth prescription medication.}

\item{atc_111a}{\link{character} ATC code of respondent's eleventh prescription medication.}

\item{atc_112a}{\link{character} ATC code of respondent's twelfth prescription medication.}

\item{atc_113a}{\link{character} ATC code of respondent's thirteenth prescription medication.}

\item{atc_114a}{\link{character} ATC code of respondent's fourteenth prescription medication.}

\item{atc_115a}{\link{character} ATC code of respondent's fifteenth prescription medication.}

\item{atc_201a}{\link{character} ATC code of respondent's first over-the-counter medication.}

\item{atc_202a}{\link{character} ATC code of respondent's second over-the-counter medication.}

\item{atc_203a}{\link{character} ATC code of respondent's third over-the-counter medication.}

\item{atc_204a}{\link{character} ATC code of respondent's fourth over-the-counter medication.}

\item{atc_205a}{\link{character} ATC code of respondent's fifth over-the-counter medication.}

\item{atc_206a}{\link{character} ATC code of respondent's sixth over-the-counter medication.}

\item{atc_207a}{\link{character} ATC code of respondent's seventh over-the-counter medication.}

\item{atc_208a}{\link{character} ATC code of respondent's eighth over-the-counter medication.}

\item{atc_209a}{\link{character} ATC code of respondent's ninth over-the-counter medication.}

\item{atc_210a}{\link{character} ATC code of respondent's tenth over-the-counter medication.}

\item{atc_211a}{\link{character} ATC code of respondent's eleventh over-the-counter medication.}

\item{atc_212a}{\link{character} ATC code of respondent's twelfth over-the-counter medication.}

\item{atc_213a}{\link{character} ATC code of respondent's thirteenth over-the-counter medication.}

\item{atc_214a}{\link{character} ATC code of respondent's fourteenth over-the-counter medication.}

\item{atc_215a}{\link{character} ATC code of respondent's fifteenth over-the-counter medication.}

\item{atc_131a}{\link{character} ATC code of respondent's first new prescription medication.}

\item{atc_132a}{\link{character} ATC code of respondent's second new prescription medication.}

\item{atc_133a}{\link{character} ATC code of respondent's third new prescription medication.}

\item{atc_134a}{\link{character} ATC code of respondent's fourth new prescription medication.}

\item{atc_135a}{\link{character} ATC code of respondent's fifth new prescription medication.}

\item{atc_231a}{\link{character} ATC code of respondent's first new over-the-counter medication.}

\item{atc_232a}{\link{character} ATC code of respondent's second new over-the-counter medication.}

\item{atc_233a}{\link{character} ATC code of respondent's third new over-the-counter medication.}

\item{atc_234a}{\link{character} ATC code of respondent's fourth new over-the-counter medication.}

\item{atc_235a}{\link{character} ATC code of respondent's fifth new over-the-counter medication.}

\item{mhr_101b}{\link{integer} Response for when the first prescription medication was last taken (1 = Today, …, 6 = Never).}

\item{mhr_102b}{\link{integer} Response for when the second prescription medication was last taken (1-6).}

\item{mhr_103b}{\link{integer} Response for when the third prescription medication was last taken (1-6).}

\item{mhr_104b}{\link{integer} Response for when the fourth prescription medication was last taken (1-6).}

\item{mhr_105b}{\link{integer} Response for when the fifth prescription medication was last taken (1-6).}

\item{mhr_106b}{\link{integer} Response for when the sixth prescription medication was last taken (1-6).}

\item{mhr_107b}{\link{integer} Response for when the seventh prescription medication was last taken (1-6).}

\item{mhr_108b}{\link{integer} Response for when the eighth prescription medication was last taken (1-6).}

\item{mhr_109b}{\link{integer} Response for when the ninth prescription medication was last taken (1-6).}

\item{mhr_110b}{\link{integer} Response for when the tenth prescription medication was last taken (1-6).}

\item{mhr_111b}{\link{integer} Response for when the eleventh prescription medication was last taken (1-6).}

\item{mhr_112b}{\link{integer} Response for when the twelfth prescription medication was last taken (1-6).}

\item{mhr_113b}{\link{integer} Response for when the thirteenth prescription medication was last taken (1-6).}

\item{mhr_114b}{\link{integer} Response for when the fourteenth prescription medication was last taken (1-6).}

\item{mhr_115b}{\link{integer} Response for when the fifteenth prescription medication was last taken (1-6).}

\item{mhr_201b}{\link{integer} Response for when the first over-the-counter medication was last taken (1-6).}

\item{mhr_202b}{\link{integer} Response for when the second over-the-counter medication was last taken (1-6).}

\item{mhr_203b}{\link{integer} Response for when the third over-the-counter medication was last taken (1-6).}

\item{mhr_204b}{\link{integer} Response for when the fourth over-the-counter medication was last taken (1-6).}

\item{mhr_205b}{\link{integer} Response for when the fifth over-the-counter medication was last taken (1-6).}

\item{mhr_206b}{\link{integer} Response for when the sixth over-the-counter medication was last taken (1-6).}

\item{mhr_207b}{\link{integer} Response for when the seventh over-the-counter medication was last taken (1-6).}

\item{mhr_208b}{\link{integer} Response for when the eighth over-the-counter medication was last taken (1-6).}

\item{mhr_209b}{\link{integer} Response for when the ninth over-the-counter medication was last taken (1-6).}

\item{mhr_210b}{\link{integer} Response for when the tenth over-the-counter medication was last taken (1-6).}

\item{mhr_211b}{\link{integer} Response for when the eleventh over-the-counter medication was last taken (1-6).}

\item{mhr_212b}{\link{integer} Response for when the twelfth over-the-counter medication was last taken (1-6).}

\item{mhr_213b}{\link{integer} Response for when the thirteenth over-the-counter medication was last taken (1-6).}

\item{mhr_214b}{\link{integer} Response for when the fourteenth over-the-counter medication was last taken (1-6).}

\item{mhr_215b}{\link{integer} Response for when the fifteenth over-the-counter medication was last taken (1-6).}

\item{mhr_131b}{\link{integer} Response for when the first new prescription medication was last taken (1-6).}

\item{mhr_132b}{\link{integer} Response for when the second new prescription medication was last taken (1-6).}

\item{mhr_133b}{\link{integer} Response for when the third new prescription medication was last taken (1-6).}

\item{mhr_134b}{\link{integer} Response for when the fourth new prescription medication was last taken (1-6).}

\item{mhr_135b}{\link{integer} Response for when the fifth new prescription medication was last taken (1-6).}

\item{mhr_231b}{\link{integer} Response for when the first new over-the-counter medication was last taken (1-6).}

\item{mhr_232b}{\link{integer} Response for when the second new over-the-counter medication was last taken (1-6).}

\item{mhr_233b}{\link{integer} Response for when the third new over-the-counter medication was last taken (1-6).}

\item{mhr_234b}{\link{integer} Response for when the fourth new over-the-counter medication was last taken (1-6).}

\item{mhr_235b}{\link{integer} Response for when the fifth new over-the-counter medication was last taken (1-6).}
}
\value{
\link{numeric} Returns 1 if the person is taking calcium channel blockers, 0 otherwise. If all medication information is missing, it returns a tagged NA.
}
\description{
This function checks if a person is taking calcium channel blockers based on the provided Anatomical Therapeutic Chemical (ATC) codes for medications
and the Canadian Health Measures Survey (CHMS) response for the time when the medication was last taken.
}
\details{
The function identifies calcium channel blockers based on ATC codes starting with "C08". It checks all medication variables provided in the input data frame.

\if{html}{\out{<div class="sourceCode">}}\preformatted{     **Missing Data Codes:**
     - The function handles tagged NAs from the `is_calcium_channel_blocker` function and propagates them.
}\if{html}{\out{</div>}}
}
\examples{
# This is a wrapper function and is not intended to be called directly by the user.
# See `is_calcium_channel_blocker` for usage examples.
}
\seealso{
\code{is_calcium_channel_blocker}
}
