% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{minperday_to_minperweek}
\alias{minperday_to_minperweek}
\title{Minutes per week from minutes per day}
\usage{
minperday_to_minperweek(MVPA_min)
}
\arguments{
\item{MVPA_min}{\link{numeric} A numeric representing the average minutes of exercise per day across a week of accelerometer use.}
}
\value{
\link{numeric} The average minutes of exercise per one week of accelerometer use.
}
\description{
This function takes the average minutes of exercise per day across a week of accelerometer use as an input (\code{MVPA_min}) and
calculates the equivalent minutes of exercise per one week of accelerometer use. The result is returned as a numeric value. This function supports vector operations.
}
\details{
The function simply multiplies the average minutes of exercise per day (\code{MVPA_min}) by 7 to obtain the equivalent
minutes of exercise per one week of accelerometer use.
}
\examples{
# Scalar usage: Single respondent
# Example: Convert average minutes of exercise per day to minutes per week.
minperday_to_minperweek(35)
# Output: 245 (The equivalent minutes of exercise per one week is 245 minutes.)

# Vector usage: Multiple respondents
minperday_to_minperweek(c(35, 40, 20))
# Returns: c(245, 280, 140)

# Database usage: Applied to survey datasets
library(dplyr)
# dataset \%>\%
#   mutate(min_per_week = minperday_to_minperweek(avg_exercise))

}
