% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variable_details}
\alias{variable_details}
\title{variable-details.csv}
\source{
See \url{https://big-life-lab.github.io/chmsflow/articles/variable_details.html} for more details.
}
\usage{
data(variable_details)
}
\description{
This dataset provides details on how variables are recoded in chmsflow.
}
\examples{
data(variable_details)
str(variable_details)
}
\keyword{datasets}
