% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycle2}
\alias{cycle2}
\title{Canadian Health Measures Survey (CHMS) Cycle 2}
\source{
Statistics Canada
}
\usage{
data(cycle2)
}
\description{
This is dummy data representing the second cycle of the Canadian
Health Measures Survey (CHMS). The CHMS survey is conducted by
Statistics Canada.
}
\examples{
data(cycle2)
str(cycle2)
}
\keyword{datasets}
