% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycle2}
\alias{cycle2}
\title{CHMS Cycle 2}
\value{
\item{cycle2}{a data frame}
}
\description{
This is synthetic data representing the second cycle of the Canadian
Health Measures Survey (CHMS). The CHMS survey is conducted by
Statistics Canada.
}
\examples{
data(cycle2)
str(cycle2)
}
\keyword{datasets}
