% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cycle2_meds}
\alias{cycle2_meds}
\title{CHMS Cycle 2 Medications}
\description{
This synthetic data representing the medication portion of the
second cycle of the Canadian Health Measures Survey (CHMS).
The CHMS survey is conducted by Statistics Canada.
}
