% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycle2_meds}
\alias{cycle2_meds}
\title{CHMS Cycle 2 Medications}
\value{
\item{cycle2_meds}{a data frame}
}
\description{
This dummy data representing the medication portion of the
second cycle of the Canadian Health Measures Survey (CHMS).
The CHMS survey is conducted by Statistics Canada.
}
\examples{
data(cycle2_meds)
str(cycle2_meds)
}
\keyword{datasets}
