% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycle3}
\alias{cycle3}
\title{CHMS Cycle 3}
\value{
\item{cycle3}{a data frame}
}
\description{
This is synthetic data representing the third cycle of the Canadian
Health Measures Survey (CHMS). The CHMS survey is conducted by
Statistics Canada.
}
\examples{
data(cycle3)
str(cycle3)
}
\keyword{datasets}
