% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycle3}
\alias{cycle3}
\title{Canadian Health Measures Survey (CHMS) Cycle 3}
\format{
A data frame with X rows and Y columns.
}
\source{
Statistics Canada
}
\usage{
data(cycle3)
}
\description{
This is dummy data representing the third cycle of the Canadian
Health Measures Survey (CHMS). The CHMS survey is conducted by
Statistics Canada.
}
\examples{
data(cycle3)
str(cycle3)
}
\keyword{datasets}
