% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycle3_meds}
\alias{cycle3_meds}
\title{CHMS Cycle 3 Medications}
\value{
\item{cycle3_meds}{a data frame}
}
\description{
This dummy data representing the medication portion of the
third cycle of the Canadian Health Measures Survey (CHMS).
The CHMS survey is conducted by Statistics Canada.
}
\examples{
data(cycle3_meds)
str(cycle3_meds)
}
\keyword{datasets}
