% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycle4}
\alias{cycle4}
\title{Canadian Health Measures Survey (CHMS) Cycle 4}
\format{
A data frame with X rows and Y columns.
}
\source{
Statistics Canada
}
\usage{
data(cycle4)
}
\description{
This is dummy data representing the fourth cycle of the Canadian
Health Measures Survey (CHMS). The CHMS survey is conducted by
Statistics Canada.
}
\examples{
data(cycle4)
str(cycle4)
}
\keyword{datasets}
