% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cycle5}
\alias{cycle5}
\title{CHMS Cycle 5}
\description{
This is synthetic data representing the fifth cycle of the Canadian
Health Measures Survey (CHMS). The CHMS survey is conducted by
Statistics Canada.
}
