% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycle5}
\alias{cycle5}
\title{CHMS Cycle 5}
\value{
\item{cycle5}{a data frame}
}
\description{
This is dummy data representing the fifth cycle of the Canadian
Health Measures Survey (CHMS). The CHMS survey is conducted by
Statistics Canada.
}
\examples{
data(cycle5)
str(cycle5)
}
\keyword{datasets}
