% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diabetes.R
\name{determine_inclusive_diabetes}
\alias{determine_inclusive_diabetes}
\title{Determine Inclusive Diabetes Status}
\usage{
determine_inclusive_diabetes(diab_m, CCC_51, diab_drug2)
}
\arguments{
\item{diab_m}{An integer indicating whether the respondent has diabetes based on HbA1c level. 1 for "Yes", 2 for "No".}

\item{CCC_51}{An integer indicating whether the respondent self-reported diabetes. 1 for "Yes", 2 for "No".}

\item{diab_drug2}{An integer indicating whether the respondent is on diabetes medication. 1 for "Yes", 0 for "No".}
}
\value{
An integer indicating the inclusive diabetes status:
- 1 ("Yes") if any of \code{diab_m}, \code{CCC_51}, or \code{diab_drug2} is 1.
- 2 ("No") if all of \code{diab_m}, \code{CCC_51}, and \code{diab_drug2} are 2.
- \code{haven::tagged_na("b")} if all three parameters are \code{NA}.
- If two parameters are \code{NA}, the third non-\code{NA} parameter determines the result.
- If one parameter is \code{NA}, the function checks the remaining two for a decision.
}
\description{
This function evaluates inclusive diabetes status based on three factors: \code{diab_m}, \code{CCC_51}, and \code{diab_drug2}.
}
\examples{

# Example: Determine the inclusive diabetes status for a respondent with diabetes based on HbA1c.
determine_inclusive_diabetes(diab_m = 1, CCC_51 = 2, diab_drug2 = 2)
# Output: 1 (Inclusive diabetes status is "Yes").

# Example: Determine the inclusive diabetes status for a respondent with all factors indicating no diabetes.
determine_inclusive_diabetes(diab_m = 2, CCC_51 = 2, diab_drug2 = 2)
# Output: 2 (Inclusive diabetes status is "No").

# Example: Determine inclusive diabetes status when only one parameter is NA.
determine_inclusive_diabetes(diab_m = 2, CCC_51 = NA, diab_drug2 = 1)
# Output: 1 (Based on `diab_drug2`, inclusive diabetes status is "Yes").

# Example: Determine inclusive diabetes status when two parameters are NA.
determine_inclusive_diabetes(diab_m = NA, CCC_51 = NA, diab_drug2 = 2)
# Output: 2 (Based on `diab_drug2`, inclusive diabetes status is "No").

}
