% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diet.R
\name{find_totalFV_cycles3to6}
\alias{find_totalFV_cycles3to6}
\title{Daily fruit and vegetable consumption in a year - cycles 3-6}
\usage{
find_totalFV_cycles3to6(
  WSDD34Y,
  WSDD35Y,
  GFVD17AY,
  GFVD17BY,
  GFVD17CY,
  GFVD17DY,
  GFVD18Y,
  GFVD19Y,
  GFVD20Y,
  GFVD22Y,
  GFVD23Y
)
}
\arguments{
\item{WSDD34Y}{\link{numeric} A numeric vector representing the number of times per year orange or grapefruit juice was consumed.}

\item{WSDD35Y}{\link{numeric} A numeric vector representing the number of times per year other fruit juices were consumed.}

\item{GFVD17AY}{\link{numeric} A numeric vector representing the number of times per year citrus fruits were consumed.}

\item{GFVD17BY}{\link{numeric} A numeric vector representing the number of times per year strawberries were consumed (in summer).}

\item{GFVD17CY}{\link{numeric} A numeric vector representing the number of times per year strawberries were consumed (outside summer).}

\item{GFVD17DY}{\link{numeric} A numeric vector representing the number of times per year other fruits were consumed.}

\item{GFVD18Y}{\link{numeric} A numeric vector representing the number of times per year tomato or tomato sauce was consumed.}

\item{GFVD19Y}{\link{numeric} A numeric vector representing the number of times per year lettuce or green leafy salad was consumed.}

\item{GFVD20Y}{\link{numeric} A numeric vector representing the number of times per year spinach, mustard greens, and cabbage were consumed.}

\item{GFVD22Y}{\link{numeric} A numeric vector representing the number of times per year potatoes were consumed.}

\item{GFVD23Y}{\link{numeric} A numeric vector representing the number of times per year other vegetables were consumed.}
}
\value{
\link{numeric} The average times per day fruits and vegetables were consumed in a year.
}
\description{
This function calculates the daily fruit and vegetable consumption in a year for respondents in the Canadian Health Measures
Survey (CHMS) cycles 3-6. It takes eleven parameters, each representing the number of times per year a specific fruit or
vegetable item was consumed. The function then sums up the consumption frequencies of all these items and divides the total
by 365 to obtain the average daily consumption of fruits and vegetables in a year.
}
\details{
The function calculates the total consumption of fruits and vegetables in a year by summing up the consumption
frequencies of all the input items. It then divides the total by 365 to obtain the average daily consumption of
fruits and vegetables in a year. NA(b) is only returned if all the parameters are missing or if the average ends
up being NA.
}
\examples{
# Scalar usage: Single respondent
# Example: Calculate average daily fruit and vegetable consumption for a cycle 3-6 respondent
# Let's assume the following annual consumption frequencies for each item:
# WSDD34Y (orange or grapefruit juice) = 50 times
# WSDD35Y (other fruit juices) = 100 times
# GFVD17AY (citrus fruits) = 150 times
# GFVD17BY (strawberries in summer) = 80 times
# GFVD17CY (strawberries outside summer) = 40 times
# GFVD17DY (other fruits) = 200 times
# GFVD18Y (tomato or tomato sauce) = 100 times
# GFVD19Y (lettuce or green leafy salad) = 80 times
# GFVD20Y (spinach, mustard greens, and cabbage) = 60 times
# GFVD22Y (potatoes) = 120 times
# GFVD23Y (other vegetables) = 90 times
# Using the function:
find_totalFV_cycles3to6(
  WSDD34Y = 50, WSDD35Y = 100, GFVD17AY = 150, GFVD17BY = 80, GFVD17CY = 40,
  GFVD17DY = 200, GFVD18Y = 100, GFVD19Y = 80, GFVD20Y = 60, GFVD22Y = 120, GFVD23Y = 90
)
# Output: 2.931507

# Example: Respondent has non-response values for all inputs.
find_totalFV_cycles3to6(
  WSDD34Y = 9998, WSDD35Y = 9998, GFVD17AY = 9998, GFVD17BY = 9998, GFVD17CY = 9998,
  GFVD17DY = 9998, GFVD18Y = 9998, GFVD19Y = 9998, GFVD20Y = 9998, GFVD22Y = 9998, GFVD23Y = 9998
)
# Output: NA

# Multiple respondents
find_totalFV_cycles3to6(
  WSDD34Y = c(50, 60), WSDD35Y = c(100, 110), GFVD17AY = c(150, 160), GFVD17BY = c(80, 90),
  GFVD17CY = c(40, 50), GFVD17DY = c(200, 210), GFVD18Y = c(100, 110), GFVD19Y = c(80, 90),
  GFVD20Y = c(60, 70), GFVD22Y = c(120, 130), GFVD23Y = c(90, 100)
)
# Returns: c(2.931507, 3.232877)

}
