% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medications.R
\name{is_NSAID}
\alias{is_NSAID}
\title{Non-steroidal anti-inflammatory drugs (NSAIDs)}
\usage{
is_NSAID(MEUCATC, NPI_25B)
}
\arguments{
\item{MEUCATC}{\link{character} ATC code of the medication.}

\item{NPI_25B}{\link{integer} Time when the medication was last taken.}
}
\value{
\link{numeric} 1 if medication is an NSAID, 0 otherwise. If inputs are invalid or out of bounds, the function returns a tagged NA.
}
\description{
This function checks if a given medication is an NSAID.
This function processes multiple inputs efficiently.
}
\details{
Identifies NSAIDs based on ATC codes starting with "M01A".

\if{html}{\out{<div class="sourceCode">}}\preformatted{     **Missing Data Codes:**
     - `MEUCATC`: `9999996` (Not applicable), `9999997-9999999` (Missing)
     - `NPI_25B`: `6` (Not applicable), `7-9` (Missing)
}\if{html}{\out{</div>}}
}
\examples{
# Scalar usage: Single respondent
is_NSAID("M01AB05", 1)
# Returns: 1

# Example: Respondent has non-response values for all inputs.
result <- is_NSAID("9999998", 8)
result # Shows: NA
haven::is_tagged_na(result, "b") # Shows: TRUE (confirms it's tagged NA(b))
format(result, tag = TRUE) # Shows: "NA(b)" (displays the tag)

# Multiple respondents
is_NSAID(c("M01AB05", "A10BB09"), c(1, 3))
# Returns: c(1, 0)

}
