% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medications.R
\name{is_other_antiHTN_med}
\alias{is_other_antiHTN_med}
\title{Determine if a CHMS respondent's medication is another anti-hypertensive drug.}
\usage{
is_other_antiHTN_med(MEUCATC, NPI_25B)
}
\arguments{
\item{MEUCATC}{A character vector representing the Anatomical Therapeutic Chemical (ATC) code of the medication.}

\item{NPI_25B}{An integer representing the CHMS response for the time when the medication was last taken.
1 = Today, 2 = Yesterday, 3 = Within the last week, 4 = Within the last month,
5 = More than a month ago, 6 = Never taken}
}
\value{
A numeric, 1 if medication is in another anti-hypertensive drug class and 0 if it is not.
}
\description{
This function checks if a given medication for a CHMS respondent belongs to another anti-hypertensive drug class.
The identification is based on the Anatomical Therapeutic Chemical (ATC) code of the medication and the time when the
medication was last taken.
}
\details{
This function uses the \code{startsWith} function to identify other anti-hypertensive drugs based on their ATC
codes, which typically start with "C02". The sub-code 'C02KX01' is excluded from the class. If the ATC code
matches the pattern and is not in the exclusion list, and the medication was taken within the last month
(NPI_25B <= 4), the medication is considered another anti-hypertensive drug, and the function returns
TRUE. Otherwise, it returns FALSE.
}
\examples{

# Let's say a for one respondent's medication, the ATC code is C02AC04 and the time they last took it was within the last 
# week. By passing in the code and integer time response (3 in this case) into the function, it can be checked if the 
# given respondent's medication is another anti-hypertensive drug (1) or not (0).

is_other_antiHTN_med(`C02AC04`, 3)

}
