% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\name{pack_years_fun}
\alias{pack_years_fun}
\title{Smoking pack-years}
\usage{
pack_years_fun(
  SMKDSTY,
  CLC_AGE,
  SMK_54,
  SMK_52,
  SMK_31,
  SMK_41,
  SMK_53,
  SMK_42,
  SMK_21,
  SMK_11
)
}
\arguments{
\item{SMKDSTY}{An integer representing the smoking status of the respondent:
\itemize{
\item 1: Daily smoker
\item 2: Occasional smoker (former daily)
\item 3: Occasional smoker (never daily)
\item 4: Former daily smoker (non-smoker now)
\item 5: Former occasional smoker (non-smoker now) who smoked at least 100 cigarettes in their lifetime
\item 6: Non-smoker (never smoked more than 100 cigarettes)
}}

\item{CLC_AGE}{A numeric value representing the respondent's age.}

\item{SMK_54}{A numeric value representing the respondent's age when they stopped smoking daily.}

\item{SMK_52}{A numeric value representing the respondent's age when they first started smoking daily.}

\item{SMK_31}{An numeric representing the number of cigarettes smoked per day for daily smokers.}

\item{SMK_41}{A numeric value representing the number of cigarettes smoked per day for occasional smokers.}

\item{SMK_53}{A numeric value representing the number of cigarettes smoked per day for former daily smokers.}

\item{SMK_42}{A numeric value representing the number of days in past month the respondent smoked at least 1 cigarette (for occasional smokers).}

\item{SMK_21}{A numeric value representing the respondent's age when they first started smoking occasionally.}

\item{SMK_11}{An integer representing whether the respondent has smoked at least 100 cigarettes in their lifetime:
\itemize{
\item 1: Yes
\item 2: No
}}
}
\value{
A numeric value representing the pack years for the respondent's smoking history.
\itemize{
\item If \code{CLC_AGE} is missing or negative, returns \code{tagged_na("b")}.
\item For different smoking statuses (\code{SMKDSTY}), the function calculates pack years as follows:
\itemize{
\item \strong{Daily smoker (1):} \code{pmax(((CLC_AGE - SMK_52) * (SMK_31 / 20)), 0.0137)}
\item \strong{Occasional smoker (former daily) (2):}
\code{pmax(((CLC_AGE - SMK_52 - (CLC_AGE - SMK_54)) * (SMK_53 / 20)), 0.0137) + ((pmax((SMK_41 * SMK_42 / 30), 1) / 20) * (CLC_AGE - SMK_54))}
\item \strong{Occasional smoker (never daily) (3):}
\code{(pmax((SMK_41 * SMK_42 / 30), 1) / 20) * (CLC_AGE - SMK_21)}
\item \strong{Former daily smoker (4):}
\code{pmax(((CLC_AGE - SMK_52 - (CLC_AGE - SMK_54)) * (SMK_53 / 20)), 0.0137)}
\item \strong{Former occasional smoker (5)}:
\itemize{
\item If \code{SMK_11 == 1} (≥100 cigarettes): \code{0.0137}
\item If \code{SMK_11 == 2} (<100 cigarettes): \code{0.007}
}
\item \strong{Non-smoker (6):} \code{0}
\item If \code{SMKDSTY} is \code{NA(a)}, returns \code{tagged_na("a")}.
\item For all other unexpected inputs, returns \code{tagged_na("b")}.
}
}
}
\description{
This function calculates an individual's smoking pack-years based on various CHMS smoking variables. Pack years is a measure used by researchers to quantify lifetime exposure to cigarette use.
}
\examples{

# Example 1: Age = 40, daily smoker, started smoking at 20, and smokes 30 cigs/day (1.5 packs/day).
pack_years_fun(SMKDSTY = 1, CLC_AGE = 40, SMK_52 = 20, SMK_31 = 30)
# Output: 30 (pack years)

# Example 2: A former occasional smoker who smoked at least 100 cigarettes in their lifetime.
pack_years_fun(
  SMKDSTY = 5, CLC_AGE = 50, SMK_54 = 40, SMK_52 = 18, SMK_31 = NA,
  SMK_41 = 15, SMK_53 = NA, SMK_42 = 3, SMK_21 = 25, SMK_11 = 1
)
# Output: 0.0137 (pack years)

}
\seealso{
https://big-life-lab.github.io/cchsflow/reference/pack_years_fun.html
}
