% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variables}
\alias{variables}
\title{List of variables in chmsflow}
\format{
A data frame with X rows and Y columns.
}
\source{
See \url{https://big-life-lab.github.io/chmsflow/articles/variables_sheet.html} for more details.
}
\usage{
data(variables)
}
\description{
This dataset lists all the variables that are present in chmsflow.
}
\examples{
data(variables)
str(variables)
}
\keyword{datasets}
